+++ apps/api/src/app/portfolio/calculator/portfolio-calculator.ts
@@ -1208,8 +1208,12 @@ export abstract class PortfolioCalculator {
 
     let cachedPortfolioSnapshot: PortfolioSnapshot | undefined;
     let isCachedPortfolioSnapshotExpired = false;
+
+    const calculationType = this.getPerformanceCalculationType();
+
     const portfolioSnapshotKey = this.redisCacheService.getPortfolioSnapshotKey(
       {
+        calculationType,
         filters: this.filters,
         userId: this.userId
       }
@@ -1248,7 +1252,7 @@ export abstract class PortfolioCalculator {
         // Compute in the background
         this.portfolioSnapshotService.addJobToQueue({
           data: {
-            calculationType: this.getPerformanceCalculationType(),
+            calculationType,
             filters: this.filters,
             userCurrency: this.currency,
             userId: this.userId
@@ -1271,7 +1275,7 @@ export abstract class PortfolioCalculator {
       // Wait for computation
       await this.portfolioSnapshotService.addJobToQueue({
         data: {
-          calculationType: this.getPerformanceCalculationType(),
+          calculationType,
           filters: this.filters,
           userCurrency: this.currency,
           userId: this.userId